requirements.txt
django
djangorestframework
django-widget-tweaks
requests
django-axes
psycopg2 # sudo apt install python3-dev libpq-dev for linux
pillow
firebase_admin